moving tests to new folder
