+++ packages/destination-actions/src/destinations/facebook-conversions-api/__tests__/flag-on-tests/pageView2.test.ts
@@ -1,7 +1,7 @@
 import nock from 'nock'
 import { createTestEvent, createTestIntegration } from '@segment/actions-core'
-import Destination from '../index'
-import { API_VERSION } from '../constants'
+import Destination from '../../index'
+import { API_VERSION } from '../../constants'
 
 const testDestination = createTestIntegration(Destination)
 const settings = {
@@ -17,7 +17,49 @@ const settingsWithTestEventCode = {
 const features = { FB_CAPI_REFACTOR_PAGE_VIEW_EVENT: true }
 
 describe('FacebookConversionsApi', () => {
-  describe('PageView', () => {
+  describe('PageView2', () => {
+    it('should throw an error syncMode incorrect', async () => {
+      nock(`https://graph.facebook.com/v${API_VERSION}/${settings.pixelId}`).post(`/events`).reply(201, {})
+
+      const event = createTestEvent({
+        event: 'Product Added',
+        userId: 'abc123',
+        properties: {
+          timestamp: 1631210030,
+          action_source: 'email',
+          email: 'test@test.com'
+        }
+      })
+
+      await expect(
+        testDestination.testAction('pageView2', {
+          event,
+          settings,
+          features,
+          mapping: {
+            __segment_internal_sync_mode: 'update',
+            currency: {
+              '@path': '$.properties.currency'
+            },
+            value: {
+              '@path': '$.properties.value'
+            },
+            action_source: {
+              '@path': '$.properties.action_source'
+            },
+            event_time: {
+              '@path': '$.properties.timestamp'
+            },
+            user_data: {
+              email: {
+                '@path': '$.properties.email'
+              }
+            }
+          }
+        })
+      ).rejects.toThrowError('Sync mode update is not supported')
+    })
+
     it('should handle a basic event', async () => {
       nock(`https://graph.facebook.com/v${API_VERSION}/${settings.pixelId}`).post(`/events`).reply(201, {})
 
@@ -33,12 +75,12 @@ describe('FacebookConversionsApi', () => {
         }
       })
 
-      const responses = await testDestination.testAction('pageView', {
+      const responses = await testDestination.testAction('pageView2', {
         event,
         settings,
         features,
         useDefaultMappings: true,
-        mapping: { action_source: { '@path': '$.properties.action_source' } }
+        mapping: { __segment_internal_sync_mode: 'add', action_source: { '@path': '$.properties.action_source' } }
       })
 
       expect(responses.length).toBe(1)
@@ -63,11 +105,12 @@ describe('FacebookConversionsApi', () => {
       })
 
       await expect(
-        testDestination.testAction('pageView', {
+        testDestination.testAction('pageView2', {
           event,
           settings,
           features,
           mapping: {
+            __segment_internal_sync_mode: 'add',
             action_source: {
               '@path': '$.properties.action_source'
             },
@@ -100,12 +143,13 @@ describe('FacebookConversionsApi', () => {
         }
       })
 
-      const responses = await testDestination.testAction('pageView', {
+      const responses = await testDestination.testAction('pageView2', {
         event,
         settings,
         features,
         useDefaultMappings: true,
         mapping: {
+          __segment_internal_sync_mode: 'add',
           action_source: { '@path': '$.properties.action_source' },
           user_data: {
             email: {
@@ -142,11 +186,12 @@ describe('FacebookConversionsApi', () => {
       })
 
       await expect(
-        testDestination.testAction('pageView', {
+        testDestination.testAction('pageView2', {
           event,
           settings,
           features,
           mapping: {
+            __segment_internal_sync_mode: 'add',
             currency: {
               '@path': '$.properties.currency'
             },
@@ -182,12 +227,12 @@ describe('FacebookConversionsApi', () => {
         }
       })
 
-      const responses = await testDestination.testAction('pageView', {
+      const responses = await testDestination.testAction('pageView2', {
         event,
         settings: settingsWithTestEventCode,
         features,
         useDefaultMappings: true,
-        mapping: { action_source: { '@path': '$.properties.action_source' } }
+        mapping: { __segment_internal_sync_mode: 'add', action_source: { '@path': '$.properties.action_source' } }
       })
 
       expect(responses.length).toBe(1)
@@ -214,12 +259,13 @@ describe('FacebookConversionsApi', () => {
         }
       })
 
-      const responses = await testDestination.testAction('pageView', {
+      const responses = await testDestination.testAction('pageView2', {
         event,
         settings: settingsWithTestEventCode,
         features,
         useDefaultMappings: true,
         mapping: {
+          __segment_internal_sync_mode: 'add',
           action_source: { '@path': '$.properties.action_source' },
           test_event_code: {
             '@path': '$.properties.test_event_code'
@@ -235,7 +281,7 @@ describe('FacebookConversionsApi', () => {
       )
     })
 
-    it('should handle a basic  event with multiple external Ids', async () => {
+    it('should handle a basic event with multiple external Ids', async () => {
       nock(`https://graph.facebook.com/v${API_VERSION}/${settings.pixelId}`).post(`/events`).reply(201, {})
 
       const event = createTestEvent({
@@ -251,12 +297,13 @@ describe('FacebookConversionsApi', () => {
         }
       })
 
-      const responses = await testDestination.testAction('pageView', {
+      const responses = await testDestination.testAction('pageView2', {
         event,
         settings,
         features,
         useDefaultMappings: true,
         mapping: {
+          __segment_internal_sync_mode: 'add',
           action_source: { '@path': '$.properties.action_source' },
           user_data: {
             externalId: {
